chore: 신고 컴포넌트 폴더 구조 리팩토링 (chat/review 도메인 기준 분리)
